+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added fuel / payload properties to SimBrief API fetcher and `main.mcf` export
+- Added fuel / payload to `tmc` export
 
 ## [2.30.0] - 2026-02-27